Enforce sorting of component props alphabetically
I believe this makes it easier to find props from a long list
and it makes the code more consistent.
There are some cases where alphabetical order might not be the
most optimal like the order of "max", "min" or with a Link
element where "params" come before "to" property.
Still I think that overall the advantages outweigh the
disadvantages.